refactor: move admin to separate folder
